backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
supabase>=2.13.0
python-dotenv>=1.1.0
httpx>=0.27.0
python-multipart>=0.0.9
PyJWT[crypto]>=2.9.0
